+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### unreleased
 - **doc:** add not supported feature list
+- **add:** transformation, linecap, linejoin and polygons
 
 ### [v0.1.1] - 2016-11-23
 - **fix:** cleanup exposed methods